requirements.txt
pygame-ce
chess
